backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow>=10.3.0
numpy>=1.26.0
opencv-python-headless>=4.8.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic>=2.10.0
slowapi>=0.1.9
